README.md
# Wi-Fi-Fingerprinting
Indoor localization for Robotics project. This app collects the Wi-Fi fingerprints in three modes: 
1. Continuous scan mode 
2. Manual scan mode 
3. Step detection scan mode. 

There are options to delete fingerprints.

First Wi-Fi accesspoint name should be entered using "Configure" option. Then select in any of the three modes mentioned above. If any of the three modes is selected, Scanning starts. It collects 10 measurements and picks the 4 best RSSIs and calculates the average and appends the fingerprint to a text file. You can stop the current scanning by clicking on the same option.